youtube_rails.rb - decouple the path matching from the domain matching
The two classes of URL - short and long - have different path
formats. But within those classes, particulary the long type, the
domain can be one of several.

Decouple the matching of the path from the matching of the domain - so
we can write regexes for each case independently, rather than
multiplying the cases and coming up with regexes which match them all.

This also allows the identifier to match the new test cases added.